requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
bcrypt==4.0.1
gunicorn==21.2.0
pytz==2025.2
requests==2.31.0

# Telegram Bot
python-telegram-bot==20.7
apscheduler==3.10.4
